Add missing tests from email.js into mjml-emails.ts, remove verifySecondary tests, delete email.js
